Final version
Hover Racing
by Mohammad Mohajerani
The game works with three players: one is the playing user, while the other two are controlled by the program via a smart way, which makes the movement extremely realistic. In the map there are ten stages multiply by three laps. In the whole of the game, there are a couple of models such as the walls, the checkpoints, the water tanks, the traps, the aisles, and the walkway models. At any stage the players should be able to pass through the checkpoint, otherwise, if they are involved in a collision with any model, they lose their acceleration as well as their health percentages and they bounce backward. If they correctly pass through the checkpoint, for three seconds the cross will be shown.  If they end up losing the 100% of their health, their vehicle is going to stop, drop on the ground and will not be able to move anymore. 
In the game dialogue the health percentages, the current speed, the current stage, the boost state, the position (rank), and the current lap, are presented in the middle. On the left of the game dialogue, the general ranking, the stage, the lap, and the health percentages of all the players are presented.
The hover-cars hover in the air like they are on some sort of gravity cushion.  All of the vehicles have particle system functionality that simulates engine fire. If any of the players make it to the last lap, the game is over and the winner and its lap time are going to be shown.
 When the game finishes the user is able to restart the game by pushing the “R” button on the keyboard. By doing so, all of the used variables will restart and the countdown will start again, indicating the beginning of the game. The same occurs at the original beginning, for 3 seconds. In order to identify the player's rank, their current laps are compared. If two or all of them are on the same lap, their current stage will be compared and if two or all of them are on the same stage, their ankle will be compared with the “dot product”, to define the rank of each one and present it on the game dialogue, which is updated every microsecond. 
In order to check the collision, different types of collision checks are applied: sphere to sphere collision, rectangle to the sphere, and a mixture of all the previous (union& intersection).
In order to control the game in the playing state, a couple of keys are used:
W=moves the car forwards
S= moves the car backward
D=steers the car clockwise
Up=moves the camera forwards
Down=moves the camera backward
Right=rotates the camera right
Left=rotates the camera left
1=resets the camera position and with the mouse, movement camera can be rotated left or right.
Space key= the boost is activated and the acceleration doubles. There is a boost warning presented and the boost itself lasts two seconds, then it overheats and it needs three seconds to reload.
2=switches the camera angle to a first-person camera view.
R=the colors of all players change randomly.